Undo accidental change

+++ BaseWeatherPlugin.mm
@@ -130,7 +130,7 @@ MSHook(void, _undimScreen, id self, SEL sel)
 
 		if (cell == nil)
 		{
-			CGFloat height = [self tableView:tableView heightForRowAtIndexPath:indexPath];
+			int height = [self tableView:tableView heightForRowAtIndexPath:indexPath];
 			cell = [[[UITableViewCell alloc] initWithFrame:CGRectMake(0, 0, [UIScreen mainScreen].bounds.size.width, height) reuseIdentifier:@"LWCalendarCell"] autorelease];
 			
 			UIImage* marker = [UIImage li_imageWithContentsOfResolutionIndependentFile:[self.plugin.bundle pathForResource:[NSString stringWithFormat:@"%@_LIClockTodayMarker", tableView.theme.sectionIconSet] ofType:@"png"]];
